Finished components and refactored for code reuse
