+++ hooks-simple/src/components/App.tsx
@@ -1,11 +1,13 @@
 import React, {useState} from 'react';
 import ResourceList from './ResourceList';
+import UserList from './UserList';
 
 const App = () => {
     const [resource, setResource] = useState('posts');
     
     return (
         <div>
+            <UserList />
             <div>
                 <button onClick={() => setResource('posts')}>Posts</button>
                 <button onClick={() => setResource('todos')}>Todos</button>
